Only uploading in the weekend.

+++ AppVeyor/Build.ps1
@@ -206,9 +206,9 @@ function CreateSource($version)
 function CheckUpload()
 {
   $day = (Get-Date).DayOfWeek
-  if ($day -ne "Saturday")
+  if ($day -ne "Saturday" -And $day -ne "Sunday")
   {
-    Write-Host "Only uploading on Saturday."
+    Write-Host "Only uploading in the weekend."
     Remove-Item ..\Windows-Distribution\*
   }
 }
